Fixed x86_64 library compilation issues (size_t-related).

+++ lib/d/src/thrift/util/openssl/bio.d
@@ -14,7 +14,7 @@ alias void BIO;
 
 enum BIO_CTRL_FLUSH = 11;
 
-int BIO_flush(BIO* b) {
+c_long BIO_flush(BIO* b) {
   return BIO_ctrl(b, BIO_CTRL_FLUSH, 0, null);
 }
